compile quick fix

+++ Brutte force/DataGenerators/estream/ciphers/ecrypt-portable.h
@@ -56,8 +56,8 @@ typedef unsigned I32T u32;
 #endif
 
 #ifdef I64T
-typedef signed long long s64;   // Ph4r05 compile fix
-typedef unsigned long long u64; // Ph4r05 compile fix
+typedef int64_t s64;   // Ph4r05 compile fix
+typedef uint64_t u64; // Ph4r05 compile fix
 #endif
 
 /*
